# docs/swagger.yaml
basePath: /
definitions:
  models.Profile:
    properties:
      email:
        type: string
      id:
        type: integer
      password:
        type: string
      username:
        type: string
    required:
    - email
    - username
    type: object
  models.Response:
    properties:
      errors:
        type: string
      message:
        type: string
      result: {}
      success:
        type: boolean
    type: object
info:
  contact: {}
  description: This is a sample server call server.
  title: CRUD Users
  version: "1.0"
paths:
  /users:
    get:
      consumes:
      - application/json
      description: list all users
      produces:
      - application/json
      responses:
        "200":
          description: string
          schema:
            type: string
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a user from JSON input
      parameters:
      - description: User Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.Profile'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Response'
      tags:
      - createusers
  /users/{id}:
    get:
      consumes:
      - application/json
      description: list One users
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: string
          schema:
            type: string
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: update  user from JSON input
      parameters:
      - description: User Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.Profile'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Response'
      tags:
      - Update
swagger: "2.0"